Agents without the voice agent and the orchestrator for scrapping and collectors
