+++ orchestrator/agent_orchestrator.py
@@ -8,7 +8,6 @@ import asyncio
 from typing import Dict, List, Any, Optional, TypedDict
 import logging
 from datetime import datetime
-import re
 import dotenv
 dotenv.load_dotenv()
 
@@ -19,6 +18,8 @@ from langchain_core.runnables import RunnableConfig
 from agents.base_agent import BaseAgent, Task, AgentResult, TaskPriority
 from agents.api_agent import create_api_agent
 from agents.scraping_agent import create_scraping_agent
+from orchestrator_nodes import OrchestratorNodes
+
 
 class OrchestratorState(TypedDict):
     """State for the orchestrator graph"""
@@ -32,6 +33,7 @@ class OrchestratorState(TypedDict):
     error: Optional[str]
     execution_plan: List[str]
 
+
 class OrchestratorAgent(BaseAgent):
     """Orchestrator agent that coordinates API and scraping agents using LangGraph"""
     
@@ -42,19 +44,15 @@ class OrchestratorAgent(BaseAgent):
         self.api_agent = create_api_agent(config)
         self.scraping_agent = create_scraping_agent(config)
         
+        # Initialize nodes container
+        self.nodes = OrchestratorNodes(
+            api_agent=self.api_agent,
+            scraping_agent=self.scraping_agent,
+            logger=self.logger
+        )
+        
         # Build LangGraph workflow
         self.workflow = self._build_workflow()
-        
-        # Query patterns for routing
-        self.query_patterns = {
-            'stock_quotes': [r'stock\s+quote', r'current\s+price', r'stock\s+price'],
-            'earnings': [r'earnings', r'quarterly\s+results', r'earnings\s+report'],
-            'economic_indicators': [r'economic\s+indicator', r'gdp', r'inflation', r'unemployment'],
-            'market_overview': [r'market\s+overview', r'market\s+summary', r'overall\s+market'],
-            'news_analysis': [r'news', r'sentiment', r'articles', r'financial\s+news'],
-            'sec_filings': [r'sec\s+filing', r'10-k', r'10-q', r'8-k', r'annual\s+report'],
-            'comprehensive_analysis': [r'comprehensive', r'complete\s+analysis', r'full\s+analysis']
-        }
     
     def _define_capabilities(self) -> List[str]:
         """Define orchestrator capabilities"""
@@ -74,15 +72,15 @@ class OrchestratorAgent(BaseAgent):
         """Build LangGraph workflow for orchestration"""
         workflow = StateGraph(OrchestratorState)
         
-        # Add nodes
-        workflow.add_node("analyze_query", self._analyze_query_node)
-        workflow.add_node("execute_api_tasks", self._execute_api_tasks_node)
-        workflow.add_node("execute_scraping_tasks", self._execute_scraping_tasks_node)
-        workflow.add_node("combine_results", self._combine_results_node)
+        # Add nodes using the separate nodes class
+        workflow.add_node("analyze_query", self.nodes.analyze_query_node)
+        workflow.add_node("execute_api_tasks", self.nodes.execute_api_tasks_node)
+        workflow.add_node("execute_scraping_tasks", self.nodes.execute_scraping_tasks_node)
+        workflow.add_node("combine_results", self.nodes.combine_results_node)
         
-        # Fix edges - make them sequential instead of parallel
+        # Set up workflow edges (sequential processing)
         workflow.add_edge("analyze_query", "execute_api_tasks")
-        workflow.add_edge("execute_api_tasks", "execute_scraping_tasks")  # Sequential
+        workflow.add_edge("execute_api_tasks", "execute_scraping_tasks")
         workflow.add_edge("execute_scraping_tasks", "combine_results")
         workflow.add_edge("combine_results", END)
         
@@ -185,334 +183,6 @@ class OrchestratorAgent(BaseAgent):
             'api_tasks_executed': len(state.get('api_results', [])),
             'scraping_tasks_executed': len(state.get('scraping_results', []))
         }
-    async def _analyze_query_node(self, state: OrchestratorState) -> OrchestratorState:
-        """Analyze user query and determine execution plan"""
-        query = state['user_query'].lower()
-        symbols = state['symbols']
-        
-        # Extract symbols from query if not provided
-        if not symbols:
-            symbols = self._extract_symbols_from_query(query)
-        
-        # Determine task type and parameters
-        task_type, task_parameters = self._determine_task_type(query, symbols)
-        
-        # Create execution plan
-        execution_plan = self._create_execution_plan(task_type, symbols)
-        
-        return {
-            **state,
-            'symbols': symbols,
-            'task_type': task_type,
-            'task_parameters': task_parameters,
-            'execution_plan': execution_plan
-        }
-    
-    async def _execute_api_tasks_node(self, state: OrchestratorState) -> OrchestratorState:
-        """Execute API-related tasks"""
-        task_type = state['task_type']
-        symbols = state['symbols']
-        task_parameters = state['task_parameters']
-        
-        api_results = []
-        
-        try:
-            # Always get stock data if symbols are provided
-            if symbols:
-                # Get stock quotes
-                result = await self.api_agent.execute('get_stock_quotes', {'symbols': symbols})
-                if result.success:
-                    api_results.append({
-                        'task': 'stock_quotes',
-                        'data': result.data,
-                        'metadata': result.metadata
-                    })
-                
-                # Get stock info from Yahoo Finance
-                result = await self.api_agent.execute('get_stock_info', {'symbols': symbols})
-                if result.success:
-                    api_results.append({
-                        'task': 'stock_info',
-                        'data': result.data,
-                        'metadata': result.metadata
-                    })
-            
-            # Always get economic indicators for market context
-            days_back = task_parameters.get('days_back', 30)
-            result = await self.api_agent.execute('get_economic_indicators', {'days_back': days_back})
-            if result.success:
-                api_results.append({
-                    'task': 'economic_indicators',
-                    'data': result.data,
-                    'metadata': result.metadata
-                })
-            
-            # Always get market overview for broader context
-            result = await self.api_agent.execute('get_market_overview', {})
-            if result.success:
-                api_results.append({
-                    'task': 'market_overview',
-                    'data': result.data,
-                    'metadata': result.metadata
-                })
-            
-        except Exception as e:
-            self.logger.error(f"Error executing API tasks: {e}")
-        
-        return {
-            **state,
-            'api_results': api_results
-        }
-    async def _execute_scraping_tasks_node(self, state: OrchestratorState) -> OrchestratorState:
-        """Execute scraping-related tasks"""
-        task_type = state['task_type']
-        symbols = state['symbols']
-        task_parameters = state['task_parameters']
-        
-        scraping_results = []
-        
-        try:
-            # Always get earnings data if symbols are provided
-            if symbols:
-                result = await self.scraping_agent.execute('batch_earnings_scrape', {
-                    'symbols': symbols,
-                    'include_calendar': True,
-                    'include_analysis': True,
-                    'include_reports': True
-                })
-                if result.success:
-                    scraping_results.append({
-                        'task': 'earnings_analysis',
-                        'data': result.data,
-                        'metadata': result.metadata
-                    })
-                
-                # Always get SEC filings for fundamental analysis
-                result = await self.scraping_agent.execute('batch_filings_scrape', {
-                    'symbols': symbols,
-                    'form_types': ['10-K', '10-Q', '8-K'],
-                    'include_foreign': True
-                })
-                if result.success:
-                    scraping_results.append({
-                        'task': 'sec_filings',
-                        'data': result.data,
-                        'metadata': result.metadata
-                    })
-            
-            # Always get financial news and sentiment for market context
-            hours_back = task_parameters.get('hours_back', 24)
-            result = await self.scraping_agent.execute('analyze_news_sentiment', {
-                'hours_back': hours_back,
-                'max_articles_per_source': 10
-            })
-            if result.success:
-                scraping_results.append({
-                    'task': 'news_sentiment',
-                    'data': result.data,
-                    'metadata': result.metadata
-                })
-            
-            # For comprehensive analysis or when symbols are provided, get detailed company analysis
-            if symbols and (task_type == 'comprehensive_analysis' or len(symbols) <= 3):
-                result = await self.scraping_agent.execute('comprehensive_company_analysis', {
-                    'symbols': symbols[:3]  # Limit to 3 symbols to avoid timeout
-                })
-                if result.success:
-                    scraping_results.append({
-                        'task': 'comprehensive_analysis',
-                        'data': result.data,
-                        'metadata': result.metadata
-                    })
-            
-        except Exception as e:
-            self.logger.error(f"Error executing scraping tasks: {e}")
-        
-        return {
-            **state,
-            'scraping_results': scraping_results
-        }
-    async def _combine_results_node(self, state: OrchestratorState) -> OrchestratorState:
-        """Combine results from API and scraping tasks"""
-        api_results = state['api_results']
-        scraping_results = state['scraping_results']
-        task_type = state['task_type']
-        symbols = state['symbols']
-        
-        final_results = {
-            'task_type': task_type,
-            'symbols_analyzed': symbols,
-            'timestamp': datetime.now().isoformat(),
-            'api_data': {},
-            'scraping_data': {},
-            'insights': {},
-            'summary': {}
-        }
-        
-        # Process API results
-        for result in api_results:
-            final_results['api_data'][result['task']] = {
-                'data': result['data'],
-                'metadata': result['metadata']
-            }
-        
-        # Process scraping results
-        for result in scraping_results:
-            final_results['scraping_data'][result['task']] = {
-                'data': result['data'],
-                'metadata': result['metadata']
-            }
-        
-        # Generate insights
-        final_results['insights'] = self._generate_insights(api_results, scraping_results, task_type)
-        
-        # Create summary
-        final_results['summary'] = self._create_summary(final_results)
-        
-        return {
-            **state,
-            'final_results': final_results
-        }
-    
-    def _extract_symbols_from_query(self, query: str) -> List[str]:
-        """Extract stock symbols from query text"""
-        # Skip symbol extraction for market overview queries
-        if 'market overview' in query.lower() or 'economic indicators' in query.lower():
-            return []
-        
-        # Common patterns for stock symbols
-        symbol_patterns = [
-            r'\b[A-Z]{1,5}\b',  # 1-5 uppercase letters
-            r'\b[A-Z]{1,4}\.[A-Z]{1,3}\b',  # Foreign symbols with exchange
-        ]
-        
-        symbols = []
-        for pattern in symbol_patterns:
-            matches = re.findall(pattern, query.upper())
-            symbols.extend(matches)
-        
-        # Enhanced filtering for common words
-        common_words = {
-            'THE', 'AND', 'OR', 'BUT', 'FOR', 'GET', 'API', 'SEC', 'NYSE', 'NEWS',
-            'WITH', 'MARKET', 'OVERVIEW', 'ECONOMIC', 'INDICATORS', 'SENTIMENT'
-        }
-        symbols = [s for s in symbols if s not in common_words]
-        
-        return list(set(symbols))  # Remove duplicates 
-    
-    def _determine_task_type(self, query: str, symbols: List[str]) -> tuple[str, Dict[str, Any]]:
-        """Determine task type based on query analysis"""
-        query_lower = query.lower()
-        
-        # Check patterns
-        for task_type, patterns in self.query_patterns.items():
-            for pattern in patterns:
-                if re.search(pattern, query_lower):
-                    # Set appropriate parameters based on task type
-                    parameters = {}
-                    if 'hours' in query_lower:
-                        hours_match = re.search(r'(\d+)\s*hours?', query_lower)
-                        if hours_match:
-                            parameters['hours_back'] = int(hours_match.group(1))
-                    
-                    if 'days' in query_lower:
-                        days_match = re.search(r'(\d+)\s*days?', query_lower)
-                        if days_match:
-                            parameters['days_back'] = int(days_match.group(1))
-                    
-                    return task_type, parameters
-        
-        # Default to comprehensive analysis if symbols are provided
-        if symbols:
-            return 'comprehensive_analysis', {}
-        else:
-            return 'market_overview', {}
-    
-    def _create_execution_plan(self, task_type: str, symbols: List[str]) -> List[str]:
-        """Create execution plan based on task type"""
-        plan = []
-        
-        if task_type in ['stock_quotes', 'market_overview', 'comprehensive_analysis']:
-            plan.append('Collect stock quotes and market data via API')
-            if symbols:
-                plan.append(f'Get detailed information for symbols: {", ".join(symbols)}')
-        
-        if task_type in ['economic_indicators', 'market_overview', 'comprehensive_analysis']:
-            plan.append('Fetch economic indicators from FRED API')
-        
-        if task_type in ['earnings', 'comprehensive_analysis']:
-            plan.append('Scrape earnings data and analysis')
-        
-        if task_type in ['sec_filings', 'comprehensive_analysis']:
-            plan.append('Collect SEC filings data')
-        
-        if task_type in ['news_analysis', 'market_overview', 'comprehensive_analysis']:
-            plan.append('Scrape financial news and analyze sentiment')
-        
-        plan.append('Combine and analyze all collected data')
-        
-        return plan
-    
-    def _generate_insights(self, api_results: List[Dict], scraping_results: List[Dict], task_type: str) -> Dict[str, Any]:
-        """Generate insights from combined results"""
-        insights = {
-            'data_quality': {},
-            'market_sentiment': {},
-            'key_findings': []
-        }
-        
-        # Analyze data quality
-        total_api_tasks = len(api_results)
-        total_scraping_tasks = len(scraping_results)
-        
-        insights['data_quality'] = {
-            'api_tasks_completed': total_api_tasks,
-            'scraping_tasks_completed': total_scraping_tasks,
-            'overall_success_rate': 1.0 if (total_api_tasks + total_scraping_tasks) > 0 else 0.0
-        }
-        
-        # Analyze market sentiment from news
-        for result in scraping_results:
-            if result['task'] == 'news_sentiment':
-                metadata = result.get('metadata', {})
-                insights['market_sentiment'] = {
-                    'overall_sentiment': metadata.get('overall_market_sentiment', 'neutral'),
-                    'articles_analyzed': metadata.get('articles_analyzed', 0),
-                    'positive_articles': metadata.get('positive_articles', 0),
-                    'negative_articles': metadata.get('negative_articles', 0)
-                }
-        
-        # Generate key findings
-        if task_type == 'comprehensive_analysis':
-            insights['key_findings'].append('Comprehensive analysis completed across multiple data sources')
-        
-        if any(r['task'] == 'stock_info' for r in api_results):
-            insights['key_findings'].append('Stock market data successfully collected')
-        
-        if any(r['task'] == 'earnings_analysis' for r in scraping_results):
-            insights['key_findings'].append('Earnings analysis and sentiment data obtained')
-        
-        return insights
-    
-    def _create_summary(self, final_results: Dict[str, Any]) -> Dict[str, Any]:
-        """Create executive summary of results"""
-        summary = {
-            'execution_summary': f"Successfully executed {final_results['task_type']} analysis",
-            'data_sources_used': [],
-            'symbols_processed': len(final_results['symbols_analyzed']),
-            'timestamp': final_results['timestamp']
-        }
-        
-        # Identify data sources used
-        if final_results['api_data']:
-            summary['data_sources_used'].extend(['Alpha Vantage', 'FRED', 'Yahoo Finance'])
-        
-        if final_results['scraping_data']:
-            summary['data_sources_used'].extend(['SEC EDGAR', 'Seeking Alpha', 'Financial News RSS'])
-        
-        summary['data_sources_used'] = list(set(summary['data_sources_used']))
-        
-        return summary
     
     # Convenience methods for common workflows
     async def _coordinate_data_collection(self, parameters: Dict[str, Any]) -> AgentResult:
@@ -554,6 +224,7 @@ class OrchestratorAgent(BaseAgent):
         )
         
         return base_health
+    
     def get_structured_output(self, result: AgentResult) -> Dict[str, Any]:
         """Get structured output suitable for passing to next agent"""
         if not result.success:
@@ -570,6 +241,8 @@ class OrchestratorAgent(BaseAgent):
             'summary': data.get('summary', {}),
             'metadata': result.metadata
         }
+
+
 # Utility function to create configured orchestrator agent
 def create_orchestrator_agent(config: Dict[str, Any] = None) -> OrchestratorAgent:
     """Factory function to create configured orchestrator agent"""
@@ -586,6 +259,7 @@ def create_orchestrator_agent(config: Dict[str, Any] = None) -> OrchestratorAgen
     
     return OrchestratorAgent(config=default_config)
 
+
 # Example usage and testing
 if __name__ == "__main__":
     import asyncio
@@ -600,16 +274,8 @@ if __name__ == "__main__":
         # Test different query types
         test_queries = [
             {
-                'query': 'Get stock quotes for AAPL and MSFT with recent earnings analysis',
-                'symbols': ['AAPL', 'MSFT']
-            },
-            {
-                'query': 'Market overview with economic indicators and news sentiment',
-                'symbols': []
-            },
-            {
-                'query': 'Comprehensive analysis for GOOGL including SEC filings',
-                'symbols': ['GOOGL']
+                'query': 'Comprehensive analysis for SAMSUNG including SEC filings',
+                'symbols': ['005930.KS','SSNLF']
             }
         ]
